updated Backup method

+++ bangertech_utility_arm.sh
@@ -212,10 +212,14 @@ if [ -z "$CHOICES" ]; then
       ;;
       '"Backup"')
         disk=$(lsblk)
-        diskint=$(whiptail --backtitle "The BangerTECH Utility ARM VERSION" --inputbox "which disk do you want to clone?\n\n$disk " 17 85 3>&1 1>&2 2>&3)
-        diskext=$(whiptail --backtitle "The BangerTECH Utility ARM VERSION" --inputbox "on which disk do you want to store your clone?\n\n$disk " 17 85 3>&1 1>&2 2>&3)
+        diskint=$(whiptail --backtitle "The BangerTECH Utility X86 VERSION" --inputbox "which disk do you want to clone?\n\n$disk " 17 85 3>&1 1>&2 2>&3)
+        diskext=$(whiptail --backtitle "The BangerTECH Utility X86 VERSION" --inputbox "on which disk do you want to store your clone?\n\n$disk " 17 85 3>&1 1>&2 2>&3)
+        if whiptail --backtitle "The BangerTECH Utility X86 VERSION" --title "MESSAGE" --yesno "Would you like to clone $diskint to $diskext now?" 8 82; then
         sudo dd if=/dev/$diskint of=/dev/$diskext bs=64K conv=noerror,sync status=progress
-        whiptail --backtitle "The BangerTECH Utility ARM VERSION" --title "WatchYourLAN" --msgbox "You´re save. Cloning complete." 8 82
+        whiptail --backtitle "The BangerTECH Utility X86 VERSION" --title "WatchYourLAN" --msgbox "You´re save. Cloning complete." 8 82
+        else 
+          whiptail --backtitle "The BangerTECH Utility X86 VERSION" --title "MESSAGE" --msgbox "Please redo the installation" 8 82
+        fi
       ;;
       '"shut-wake"')
         timeshutdown=$(whiptail --backtitle "The BangerTECH Utility ARM VERSION" --inputbox " when do you want to shutdown your server? (hh:mm) " 15 85 3>&1 1>&2 2>&3)
